+++ pkgs/ffigen/test/code_generator_tests/ffigen_smart_regen/headers/file.c
@@ -1,0 +1,21 @@
+#include "header1.h"
+
+int doNothingToStruct1(Color dumInput) {
+  return 0;
+}
+
+int doNothingToStruct2(Vertex dumInput) {
+  return 0;
+}
+
+int doNothingToStruct3(Vector2 dumInput) {
+  return 0;
+}
+
+int doNothingToStruct4(Vector3 dumInput) {
+  return 0;
+}
+
+int doNothingToStruct5(Vector4 dumInput) {
+  return 0;
+}